🔨 Forge: Extract Barrel File for Input Components
Extracted multiple individual imports in InputPanel.tsx and other consumers into a single barrel file (`src/components/inputs/index.ts`). This reduces top-of-file import clutter, adhering to the "Boy Scout Rule" and making the codebase cleaner and easier to read for humans. Added an ADR entry in `.jules/forge.md` recording this architectural decision.

Tested via `pnpm lint` and `pnpm test`.
