+++ .jules/forge.md
@@ -1,1 +1,6 @@
 ## 2025-04-28 - [Barrel File for Style Controls] **Context:** [Multiple individual imports in StyleControls.tsx clutters the top of the file] **Decision:** [Use a barrel file (index.ts) to re-export related components from style-controls directory] **Consequence:** [Cleaner imports in consumers, but requires maintaining the barrel file]
+
+## 2026-04-29 - [Barrel File for Input Components]
+**Context:** [Multiple individual imports in InputPanel.tsx clutters the top of the file]
+**Decision:** [Use a barrel file (index.ts) to re-export related components from inputs directory]
+**Consequence:** [Cleaner imports in consumers, but requires maintaining the barrel file]
