+++ src/components/inputs/index.ts
@@ -1,0 +1,17 @@
+export * from './EmailInput';
+export * from './EventInput';
+export * from './FormFields';
+export * from './InputRegistry';
+export * from './LocationInput';
+export * from './MeetingInput';
+export * from './PaymentInput';
+export * from './PhoneInput';
+export * from './SmsInput';
+export * from './SocialInput';
+export * from './TextInput';
+export * from './TypeSelector';
+export * from './UrlInput';
+export * from './VCardInput';
+export * from './WifiInput';
+export * from './styles';
+export * from './useInputLogic';
